+++ plat/st/stm32mp1/stm32mp1_def.h
@@ -250,4 +250,9 @@ static inline uint32_t tamp_bkpr(uint32_t idx)
  ******************************************************************************/
 #define I2C4_BASE			U(0x5C002000)
 
+/*******************************************************************************
+ * Device Tree defines
+ ******************************************************************************/
+#define DT_RCC_CLK_COMPAT		"st,stm32mp1-rcc"
+
 #endif /* STM32MP1_DEF_H */
